+++ src/extensions/session-calenar/calendar-constants/calendar-settings-menu.ts
