+++ src/extensions/session-calenar/calendar-constants/calendar-slash-commands.ts
