+++ src/extensions/session-calenar/interaction-handling/modal-handler.ts
