+++ src/extensions/session-calenar/shared-calendar-functions.ts
